+++ src/indexHelpers.ts
@@ -35,3 +35,4 @@ export { default as zip } from "./helpers/files/zip";
 export { default as exportToExcel } from "./helpers/files/exportToExcel";
 export { default as saveAsFile } from "./helpers/files/saveAsFile";
 export { csvFromData, csvToData, csvDefineDelimiter } from "./helpers/files/csv";
+export { default as imageConvert } from "./helpers/files/imageConvert";